Fix Pomodoro create alarm function so that refreshing tokens work
Otherwise, org-gcal attempts to use the global values of the `defcustom`
variables set in the `let*`, which aren't necessarily the ones that should be
used for the alarm calendar. In the future, org-gcal should be modified to
support per-calendar client IDs and secrets, and then this hack can be
removed.