Switch to local checkout of Org 9.2
Primarily I'm doing this to fix the behavior of the `:tags` clocktable option
in Org 9.1. This was fixed in Org 9.2, but the option was also renamed to
`:match`. Since Org 9.2 doesn't seem to be coming out anytime soon, and I need
this feature for my clock reports, I'm going to keep a local checkout.

As part of this, several options have to be changed:

- Locally load the org-mode.git repo (which contains the contrib file) instead
  of installing org-plus-checkout from ELPA.
- Change the templates using `:tags` to use `:match`
- Remove obsolete variable `org-babel-use-quick-and-dirty-noweb-expansion`.
- Change a bunch of `:ensure` options to `use-package` to use `:after`
  instead (which I should have done anyway).